feat(@binstruct/png)!: refactor zlib handling with bit-packed header and remove legacy code
